refactor: remove .toLong() as not needed

+++ sts-mock/token.groovy
@@ -81,7 +81,7 @@ def buildAccessToken(Map payload) {
             aud                   : issuerBaseUrl,
             credential_identifiers: payload.credential_identifiers,
             c_nonce               : UUID.randomUUID().toString(),
-            exp                   : (System.currentTimeMillis() / 1000).toLong() + 180,
+            exp                   : (System.currentTimeMillis() / 1000) + 180,
             jti                   : UUID.randomUUID().toString()
     ]
     def encodedPayload = Base64.urlEncoder.withoutPadding().encodeToString(JsonOutput.toJson(accessTokenPayload).bytes)
